Add files via upload
StimulationFile_NewStreamMaster_Fra is the code used to process and extract information from the MedAssociate files recorded
during optogenetic stimulation experiments.
This code has been tested on matlab R2018a and R2021b on a windows 10 computer.
This master code recall a function "Analysis_5minON-OFF" through a question asked in the command window.
Read carefully the notes at the beginning of the Master script to know how to setup the file and folder and subfolders
for the analysis.
The code will go into each subfolder, analyze the files within the subfolder and save a matlab variable having the same subfolder
name. This allows the processing of multiple sessions or groups automatically.

MedPC_Analysis_FR is the code used to extract the time stamps from the MedAssociate software used to record mouse behavior.
This code has been tested on matlab R2018a and R2021b on a windows 10 computer.
Extensive explanation on how to use the code can be found in the comments at the beginning of the script.
For further questions please contact the corresponding authors of the paper.

If you have further questions please contact the corresponding authors of the paper.